refactor: name functions in handle cases, and resolve #4
