+++ src/rules/canonical-vocabulary/formatMessage.spec.ts
