+++ src/rules/canonical-vocabulary/formatMessage.ts
