+++ src/rules/canonical-vocabulary/handleCases.spec.ts
