+++ src/utils/stringsAreEquivalent.spec.ts
